sliding puzzle (#1237)
Note: I put some effort toward a completely general, automated solution, but eventually just reduced the board size to 3x3 and solved it manually, and used the API to obtain the manual commands to paste into the solution:

    curl http://localhost:5357/repl/history/full | jq .[].in -r

The board generator is generalized to any square dimension, but some scaffolding must be built to re-use the code across multiple scenarios.  The new `Sliding Puzzles` directory is intended to contain a 3x3, 4x4, and 5x5 puzzle, but currently only contains 3x3.

### 3x3 board
![image](https://github.com/swarm-game/swarm/assets/261693/babf3dbb-2d89-4c1e-a452-be8f61077180)

### 4x4 demo
![board](https://user-images.githubusercontent.com/261693/235421110-db0de5ce-8a25-482f-8a61-f42f5b469761.png)

## Demo

    scripts/play.sh --scenario "data/scenarios/Challenges/Sliding Puzzles/3x3.yaml" --autoplay