WIP: do away with default, make `WorldFun` a `Monoid`
Complicated by the fact that we sometimes want to have transparency
on the "entity channel" (e.g. to stamp a structure with some
transparent parts over existing entities etc.) and sometimes want to
"erase" entities (e.g. when the default background has an entity like
water and in some places we want to override it).  Add `Erasable` type
to deal with this.