cache REPL history
I think this is the proper way to do things: we give the entire REPL
history to brick so that it can render the entire thing and thus draw
the scroll bar correctly, scroll through the history with PgUp/PgDown
etc., but we cache it so that it is only re-rendered when the history
changes (i.e. when a new input is entered, or a new output becomes
available).