invalidate cache when starting new game
This fixes an issue introduced in #1481, where the old REPL history
would still be visible when starting a new scenario until entering a
command for the first time.

Incidentally, this also fixes the issue where the previous world would be
sometimes briefly visible before showing the new world upon starting a
new scenario.