iterative deepening does not work!
We really do need a queue of locations.  We could create one now via
`list.sw` --- i.e. Gödel-encode each coordinate pair as an int, then
make a banker's queue with two lists of ints, themselves Gödel-encoded
as lists.  That sounds super tedious though, might be better to wait
until we have real recursive types.

Even with queues, getting from one location in the queue to the next
is tricky.  Have to also store the BFS tree and use it to navigate
from one to the next??