Depend only on `vty-crossplatform`
Closes #1623.

Note that we still have to list both `vty-unix` and `vty-windows` in
`stack.yaml` but that's OK; it just means they are both made available
as extra dependencies, but only whichever is needed will actually be
installed.  In the codebase itself, we now get to avoid CPP on
imports, and the code is quite a bit simpler.