refactor to remove `swarm-tournament` dep from `swarm-integration`
The only reason the `swarm-integration` test had `swarm-tournament` as
a dependency was because of a few validation functions in
`Swarm.Web.Tournament.Validate`.  However, this seemed to me like a
strange place for those validation functions to live, since they had
nothing to do with running a tournament in particular, and it meant
that running the integration tests required having postgresql
installed even though no databases are used in the integration tests
at all.

This PR moves those validation functions to two more generic places:
the first function to extract text from a sequence of log messages now
lives in `Swarm.Log`, and the other two functions to check for bad
errors and run a `GameState` to completion now live in a new module
`Swarm.Game.Step.Validate`.