Update recipe graph generation to take `yields` into account
I noticed this when regenerating the recipe graph on the wiki and saw
that `tea leaves` was sitting there with no input.  `tea leaves` are
obtained by harvesting `tea plant`s (which now occur naturally in the
world) but we were not taking such "yield" pairs into account when
creating the graph.  This PR adds a directed edge for each entity that
`yields` another.

The image is already updated on the wiki with the output from this PR:

  https://github.com/swarm-game/swarm/wiki/Recipes-cheat-sheet