rename `doc-fragments` to `_doc-fragments` (#1821)
Recently the test suite started failing for me with an error about not finding an `00-ORDER.txt` file in `doc-fragments`.  I'm not sure why it never failed before, but the reason for the failure is clear enough: any folder under `data/scenarios` which does not start with an underscore is assumed to contain scenarios.